EP-1171: mere int. testing
