+++ src/test/kotlin/no/nav/eessi/pensjon/integrasjonstest/IntegrasjonsTestConfig.kt
@@ -1,0 +1,50 @@
+package no.nav.eessi.pensjon.integrasjonstest
+
+import org.apache.kafka.clients.consumer.ConsumerConfig
+import org.apache.kafka.clients.producer.ProducerConfig
+import org.apache.kafka.common.serialization.StringDeserializer
+import org.apache.kafka.common.serialization.StringSerializer
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.boot.test.context.TestConfiguration
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Primary
+import org.springframework.kafka.core.ConsumerFactory
+import org.springframework.kafka.core.DefaultKafkaConsumerFactory
+import org.springframework.kafka.core.DefaultKafkaProducerFactory
+import org.springframework.kafka.core.KafkaTemplate
+import org.springframework.kafka.core.ProducerFactory
+import org.springframework.kafka.test.EmbeddedKafkaBroker
+
+@TestConfiguration
+class IntegrasjonsTestConfig {
+    @Value("\${" + EmbeddedKafkaBroker.SPRING_EMBEDDED_KAFKA_BROKERS + "}")
+    private lateinit var brokerAddresses: String
+
+    @Bean
+    fun producerFactory(): ProducerFactory<String, String> {
+        val configs = HashMap<String, Any>()
+        configs[ProducerConfig.BOOTSTRAP_SERVERS_CONFIG] = this.brokerAddresses
+        configs[ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG] = StringSerializer::class.java
+        configs[ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG] = StringSerializer::class.java
+        return DefaultKafkaProducerFactory(configs)
+    }
+
+    @Bean
+    fun consumerFactory(): ConsumerFactory<String, String> {
+        val configs = HashMap<String, Any>()
+        configs[ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG] = this.brokerAddresses
+        configs[ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG] = StringDeserializer::class.java
+        configs[ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG] = StringDeserializer::class.java
+        configs[ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG] = false
+        configs[ConsumerConfig.AUTO_OFFSET_RESET_CONFIG] = "earliest"
+        configs[ConsumerConfig.GROUP_ID_CONFIG] = "eessi-pensjon-group-test"
+
+        return DefaultKafkaConsumerFactory(configs)
+    }
+
+    @Bean
+    @Primary
+    fun kafkaTemplate(): KafkaTemplate<String, String> {
+        return KafkaTemplate(producerFactory())
+    }
+}
