+++ src/test/kotlin/no/nav/eessi/pensjon/integrasjonstest/mottatt/PBuc01UgyldigFNRIntegrasjonsTest.kt
@@ -2,19 +2,19 @@ package no.nav.eessi.pensjon.integrasjonstest.mottatt
 
 import no.nav.eessi.pensjon.eux.model.buc.Buc
 import no.nav.eessi.pensjon.eux.model.buc.Participant
+import no.nav.eessi.pensjon.integrasjonstest.CustomMockServer
 import no.nav.eessi.pensjon.integrasjonstest.IntegrasjonsBase
+import no.nav.eessi.pensjon.integrasjonstest.IntegrasjonsTestConfig
+import no.nav.eessi.pensjon.integrasjonstest.OPPGAVE_TOPIC
+import no.nav.eessi.pensjon.integrasjonstest.SED_MOTTATT_TOPIC
 import no.nav.eessi.pensjon.json.toJson
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.kafka.test.context.EmbeddedKafka
 import org.springframework.test.annotation.DirtiesContext
 import org.springframework.test.context.ActiveProfiles
-import java.util.concurrent.TimeUnit
 
-private const val SED_MOTTATT_TOPIC = "eessi-basis-sedMottatt-v1"
-private const val OPPGAVE_TOPIC = "eessi-pensjon-oppgave-v1"
-
-@SpringBootTest( classes = [IntegrasjonsBase.TestConfig::class], value = ["SPRING_PROFILES_ACTIVE", "integrationtest"])
+@SpringBootTest( classes = [IntegrasjonsTestConfig::class], value = ["SPRING_PROFILES_ACTIVE", "integrationtest"])
 @ActiveProfiles("integrationtest")
 @DirtiesContext
 @EmbeddedKafka(topics = [SED_MOTTATT_TOPIC, OPPGAVE_TOPIC])
@@ -23,9 +23,7 @@ internal class PBuc01UgyldigFNRIntegrasjonsTest : IntegrasjonsBase() {
     @Test
     fun `Sender Pensjon SED (P2000) med ugyldig FNR og forventer routing til 4303`() {
 
-        //given a person
-
-        //given a http service with buc and sed
+        //setup server
         CustomMockServer()
             .medJournalforing(false, "429434388")
             .medNorg2Tjeneste()
@@ -46,12 +44,11 @@ internal class PBuc01UgyldigFNRIntegrasjonsTest : IntegrasjonsBase() {
                 "/pdf/pdfResponseUtenVedlegg.json"
             )
 
-        //when receiving a p2000 with invalid fnr
+        //send msg
         initAndRunContainer(SED_MOTTATT_TOPIC, OPPGAVE_TOPIC).also {
-                it.kafkaTemplate.send(SED_MOTTATT_TOPIC, javaClass.getResource("/eux/hendelser/P_BUC_01_P2000_ugyldigFNR.json").readText())
+            it.sendMsgOnDefaultTopic("/eux/hendelser/P_BUC_01_P2000_ugyldigFNR.json")
+            it.waitForlatch(sedListener)
         }
-
-        sedListener.getSendtLatch().await(10, TimeUnit.SECONDS)
 
         //then route to 4303
         OppgaveMeldingVerification("429434388")
