+++ src/test/kotlin/no/nav/eessi/pensjon/integrasjonstest/sendt/PBuc01P2000IntegrasjonsTest.kt
@@ -2,19 +2,19 @@ package no.nav.eessi.pensjon.integrasjonstest.sendt
 
 import no.nav.eessi.pensjon.eux.model.buc.Buc
 import no.nav.eessi.pensjon.eux.model.buc.Participant
+import no.nav.eessi.pensjon.integrasjonstest.CustomMockServer
 import no.nav.eessi.pensjon.integrasjonstest.IntegrasjonsBase
+import no.nav.eessi.pensjon.integrasjonstest.IntegrasjonsTestConfig
+import no.nav.eessi.pensjon.integrasjonstest.OPPGAVE_TOPIC
+import no.nav.eessi.pensjon.integrasjonstest.SED_SENDT_TOPIC
 import no.nav.eessi.pensjon.json.toJson
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.kafka.test.context.EmbeddedKafka
 import org.springframework.test.annotation.DirtiesContext
 import org.springframework.test.context.ActiveProfiles
-import java.util.concurrent.TimeUnit
 
-private const val SED_SENDT_TOPIC = "eessi-basis-sedSendt-v1"
-private const val OPPGAVE_TOPIC = "eessi-pensjon-oppgave-v1"
-
-@SpringBootTest(classes = [IntegrasjonsBase.TestConfig::class],  value = ["SPRING_PROFILES_ACTIVE", "integrationtest"])
+@SpringBootTest(classes = [IntegrasjonsTestConfig::class],  value = ["SPRING_PROFILES_ACTIVE", "integrationtest"])
 @ActiveProfiles("integrationtest")
 @DirtiesContext
 @EmbeddedKafka(topics = [SED_SENDT_TOPIC, OPPGAVE_TOPIC])
@@ -22,6 +22,7 @@ internal class PBuc01P2000IntegrasjonsTest : IntegrasjonsBase() {
 
     @Test
     fun `Når en sedSendt hendelse blir konsumert skal det opprettes journalføringsoppgave for pensjon SEDer`() {
+        //setup server
         CustomMockServer()
             .medJournalforing(false, "429434378")
             .medNorg2Tjeneste()
@@ -40,12 +41,13 @@ internal class PBuc01P2000IntegrasjonsTest : IntegrasjonsBase() {
                 "/pdf/pdfResponseUtenVedlegg.json"
             )
 
+        //send msg
         initAndRunContainer(SED_SENDT_TOPIC, OPPGAVE_TOPIC).also {
-            it.kafkaTemplate.send(SED_SENDT_TOPIC, javaClass.getResource("/eux/hendelser/P_BUC_01_P2000.json").readText())
+            it.sendMsgOnDefaultTopic("/eux/hendelser/P_BUC_01_P2000.json")
+            it.waitForlatch(sedListener)
         }
-        sedListener.getSendtLatch().await(10, TimeUnit.SECONDS)
 
-        //then we expect automatisk journalforing
+        //verify route
         OppgaveMeldingVerification("429434378")
             .medHendelsetype("SENDT")
             .medSedtype("P2000")
