+++ src/test/kotlin/no/nav/eessi/pensjon/integrasjonstest/sendt/PBuc05X008IntegrasjonsTest.kt
@@ -2,7 +2,11 @@ package no.nav.eessi.pensjon.integrasjonstest.sendt
 
 import no.nav.eessi.pensjon.eux.model.buc.Buc
 import no.nav.eessi.pensjon.eux.model.buc.Participant
+import no.nav.eessi.pensjon.integrasjonstest.CustomMockServer
 import no.nav.eessi.pensjon.integrasjonstest.IntegrasjonsBase
+import no.nav.eessi.pensjon.integrasjonstest.IntegrasjonsTestConfig
+import no.nav.eessi.pensjon.integrasjonstest.OPPGAVE_TOPIC
+import no.nav.eessi.pensjon.integrasjonstest.SED_SENDT_TOPIC
 import no.nav.eessi.pensjon.json.toJson
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
@@ -11,11 +15,8 @@ import org.springframework.kafka.core.KafkaTemplate
 import org.springframework.kafka.test.context.EmbeddedKafka
 import org.springframework.test.annotation.DirtiesContext
 import org.springframework.test.context.ActiveProfiles
-import java.util.concurrent.TimeUnit
 
-private const val SED_SENDT_TOPIC = "eessi-basis-sedSendt-v1"
-private const val OPPGAVE_TOPIC = "eessi-pensjon-oppgave-v1"
-@SpringBootTest(classes = [IntegrasjonsBase.TestConfig::class], value = ["SPRING_PROFILES_ACTIVE", "integrationtest"])
+@SpringBootTest(classes = [IntegrasjonsTestConfig::class], value = ["SPRING_PROFILES_ACTIVE", "integrationtest"])
 @ActiveProfiles("integrationtest")
 @DirtiesContext
 @EmbeddedKafka(topics = [SED_SENDT_TOPIC, OPPGAVE_TOPIC], partitions = 1)
@@ -27,7 +28,7 @@ internal class PBuc05X008IntegrasjonsIntegrasjons : IntegrasjonsBase() {
     @Test
     fun `Når en sed (X008) hendelse blir konsumert skal det opprettes journalføringsoppgave`() {
 
-        //given a http service with buc and sed
+        //server setup
         CustomMockServer()
             .medJournalforing(false, "429434379")
             .medNorg2Tjeneste()
@@ -43,12 +44,13 @@ internal class PBuc05X008IntegrasjonsIntegrasjons : IntegrasjonsBase() {
             .medEuxGetRequest("/buc/161558/sed/44cb68f89a2f4e748934fb4722721018","/sed/P2000-ugyldigFNR-NAV.json")
             .medEuxGetRequest( "/buc/161558/sed/40b5723cd9284af6ac0581f3981f3044/filer","/pdf/pdfResonseMedP2000MedVedlegg.json" )
 
+        //send msg
         initAndRunContainer(SED_SENDT_TOPIC, OPPGAVE_TOPIC).also {
-            it.kafkaTemplate.send(SED_SENDT_TOPIC,javaClass.getResource("/eux/hendelser/P_BUC_05_X008.json").readText())
+            it.sendMsgOnDefaultTopic("/eux/hendelser/P_BUC_05_X008.json")
+            it.waitForlatch(sedListener)
         }
-        sedListener.getSendtLatch().await(10, TimeUnit.SECONDS)
 
-        //then route to 4303
+        //verify route
         OppgaveMeldingVerification("429434379")
             .medHendelsetype("SENDT")
             .medSedtype("X008")
